Revert "Merge branch 'main' into feature/editauth"
This reverts commit 7a6090f5c182f7adfc5251b08feda259605c7e50, reversing
changes made to dd6360ae6b513fd55e0e4d3c043c5651ce5c4830.
