+++ client/src/screens/getsuggestions.tsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from "react";
-import { ScrollView, StyleSheet, TextInput, Alert, Modal, Linking, TouchableOpacity } from "react-native";
+import { ScrollView, StyleSheet, TextInput, Alert, Modal } from "react-native";
 import { Text, View, Colors, Button, Checkbox } from "react-native-ui-lib";
 import { observer } from "mobx-react";
 import { NavioScreen } from "rn-navio";
 import { services, useServices } from "@app/services";
 import { useAppearance } from "@app/utils/hooks";
 import { LocationSearchApi } from "@app/services/api/locationsearch";
-import { LocationDetailsApi } from "@app/services/api/locationdetails";
 
 export const GetSuggestions: NavioScreen = observer(() => {
   useAppearance();
@@ -41,20 +40,7 @@ export const GetSuggestions: NavioScreen = observer(() => {
     try {
       const searchQuery = [location, ...selectedRecreation, ...selectedDiner].join(", ");
       const data = await LocationSearchApi.search(searchQuery, "attractions");
-      const results = data?.data || [];
-
-      const detailedResults = await Promise.all(
-        results.map(async (item: any) => {
-          try {
-            const details = await LocationDetailsApi.getDetails(Number(item.location_id));
-            return details ?? item;
-          } catch {
-            return item;
-          }
-        })
-      );
-
-      setSuggestions(detailedResults);
+      setSuggestions(data?.data || []);
     } catch (error) {
       console.error("Error fetching suggestions:", error);
     } finally {
@@ -79,28 +65,7 @@ export const GetSuggestions: NavioScreen = observer(() => {
           <Text text70M>Loading suggestions...</Text>
         ) : (
           suggestions.map((item, index) => (
-            <TouchableOpacity
-              key={index}
-              onPress={() => {
-                if (item.web_url) {
-                  Linking.openURL(item.web_url);
-                }
-              }}
-              style={{
-                padding: 16,
-                borderWidth: 1,
-                borderColor: "#ccc",
-                borderRadius: 12,
-                marginBottom: 15,
-                backgroundColor: "#fff",
-              }}
-            >
-              <Text text60BO marginB-s1>{item.name}</Text>
-              <Text text70 marginB-s1>{item.address_obj?.address_string}</Text>
-              <Text>
-                {Array.from({ length: Math.round(Number(item.rating) || 0) }, () => "⭐").join("") || "No rating"}
-              </Text>
-            </TouchableOpacity>
+            <Text key={index} text70M>{item.name}</Text>
           ))
         )}
       </ScrollView>
@@ -117,7 +82,7 @@ export const GetSuggestions: NavioScreen = observer(() => {
               onValueChange={() => toggleSelection(option, 'recreation')} 
             />
           ))}
-          <Button label="Close" marginT-s4 onPress={() => setShowRecreationModal(false)} />
+          <Button label="Submit" marginT-s4 onPress={() => setShowRecreationModal(false)} />
         </View>
       </Modal>
 
@@ -133,7 +98,7 @@ export const GetSuggestions: NavioScreen = observer(() => {
               onValueChange={() => toggleSelection(option, 'diner')} 
             />
           ))}
-          <Button label="Close" marginT-s4 onPress={() => setShowDinerModal(false)} />
+          <Button label="Submit" marginT-s4 onPress={() => setShowDinerModal(false)} />
         </View>
       </Modal>
     </View>
